\name{adults.3dHeat.logFoldChange}
\alias{adults.3dHeat.logFoldChange}
\docType{data}
\title{
Heat stress response of adult coral
}
\description{
Acropora millepora (adult) response to three days of heat stress (31.5oC)
log-fold-changes inferred using DESeq package from tag-based RNA-seq data.
}
\usage{data("adults.3dHeat.logFoldChange")}
\format{
  A data frame with 44363 observations on the following 2 variables.
  \describe{
    \item{\code{gene}}{gene id, a factor with 44363 levels }
    \item{\code{lfc}}{log fold-change, a numeric vector}
  }
}
