\name{gene2kog}
\alias{gene2kog}
\docType{data}
\title{
KOG class annotations
}
\description{
KOG class annotations for Acropora millerpoa transcriptome.
}
\usage{data("gene2kog")}
\format{
  A data frame with 16175 observations on the following 2 variables.
  \describe{
    \item{\code{V1}}{a factor with 16175 levels }
    \item{\code{V2}}{a factor with 23 levels }
  }
}
\source{
https://dl.dropboxusercontent.com/u/37523721/amillepora_transcriptome_july2014.zip
}
\references{
Transcriptome assembly: Moya et al (2012), Mol Ecol 21:2440-2454.
Transcriptome annotation: Dixon GB, Davies SW, Aglyamova GA, Meyer E, Bay LK and Matz MV (2015) Genomic determinants of coral heat tolerance across latitudes. 
}
